+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 [Unreleased](https://github.com/jshwi/readmetester/compare/v2.0.0...HEAD)
 ------------------------------------------------------------------------
 ### Added
+- Adds linter for README
 - Adds additional errors inheriting from `OutputDocumentError`
 - Adds `--version` optional argument